Updated Shape2SAS to version 11
Version 10:
Fixed a small bug with regards to rotation point causing the subunits to remove points even though no overlap existed.

version 11:
added a global call to all classes if the user inputs a class name to a subunit or a structure factor that is not in the name dictionary. This makes it such that the user may not have to interact with other aspects of the code as long as they use the same name as the class name as input to Shape2SAS

Updated Markdown file with a description of how to add structure factor and subunits to Shape2SAS + new input parameters to the table.